// backend/src/types.ts
export interface PlayerProgress {
  displayName: string | null;
  level: number;
  experience: number;
  solvedCases: number;
  failedCases: number;
  abandonedCases: number;
  nextLevelExperience: number | null;
}